pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ghostpost"
version = "0.1.0"
description = "Agent-first email management system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "google-api-python-client>=2.150",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "pyjwt>=2.9",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "click>=8.1",
    "apscheduler>=3.10",
    "websockets>=14.0",
    "httpx>=0.27",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "anthropic>=0.80",
]

[project.scripts]
ghostpost = "src.cli.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "allow_alerts_file_write: mark test as intentionally exercising _append_alert with a safe temp-dir ALERTS_FILE redirect",
]
